Summary: added bugs ratio and resolved ratio
In summary page, added two columns in by_project, by_enum, by_developer
and by_reporter summary tables:
- resolved bug ratio, i.e. (resolved + closed)/total * 100%
- item ratio, i.e. (item total)/(sum of total bugs).

A new 'summary_header' string variable was added in replacement of the
old 'orct' one, to support the 2 new columns; structure and processing
of this string was changed from the original, it is no longer wrapped in
parenthesis, so it can be parsed with a simple explode() call instead of
relying on preg_split().

This commit  is a squash of rainflood's original work, following
discussion in PR https://github.com/mantisbt/mantisbt/pull/1231,
including some adjustments by dregad.

Note: the change includes refactoring of the original code, to reduce
duplication, remove useless code and simplify it.

Fixes #23627

Signed-off-by: Damien Regad <dregad@mantisbt.org>