Fixed case-sensitive LDAP attributes names
A ldap_get_entries() function returns attribute names as lowercase keys so next call
of array_key_exists() fails because of it performs a case-sensitive comparison.
For example if a $g_ldap_realname_field option is set to 'givenName'
(a well-known LDAP attribute display name) then it finds nothing.

Fixes #24043