Catch unserialize errors in install_stored_filter_migrate()
Instead of just choking on invalid data when unserializing it (and
possibly display a PHP notice, depending on error reporting settings),
we now catch the error and display a friendly and useful error message.

This also fixes a potential error when handling invalid JSON, since
json_decode() returns null (not false) in this case.