Add filter for the “last updated“ date in the soap api
Currently the soap method `mc_filter_search_issues` allows to filter
the result set by the “sumbitted” date of the issue. But not with the
“last updated” date.

For a task, we need to get the latest updated issues, to execute some
business logic. Therefore expand the soap api, with this additional
filter, so that's now also posssible to filter the result-set of the
search with the last “updated date”.

- Add this new filter to the `mc_filter_api`, simular like the
  “submitted date”.
- Add it also the the wsdl, so that this new api is documented for
  other programms.

Fixes #24717