MantisGraph, update includes and init
Since 1.3, plugins don't need to explicitly include core apis.
The proper way to include plugin's own apis is with
plugin_require_api().
For a better organization, plugin apis can be inluded in the plugin
init() method, so it does not need to be included for each plugin's
page. An additional benefit is that these apis are only included when
the plugin is enabled, as opposed to having the includes in the plain
php file.