Fix missing boxes in my view page
In My View page, some boxes are discarded if the user don't have
certain permissions. Those checks are performed against the current
project only, however, the boxes may show issues from subprojects
where the user can have those permissions.

With this change, the checks are made for the actual list of projects
that will be included in the filters for those sections.

Fixes: #23333