Use true instead of null for $p_obsolete param
As discussed in PR https://github.com/mantisbt/mantisbt/pull/1492#discussion_r270359921
the semantics of the $p_obsolete parameter have changed, so the proper
fix for the regression (issue #25561), is to update the callers to use
`true` instead of `null`.

This commit addresses the 3 identified occurrences of such calls to
version_get_all_rows().