Fix PHP Notice in user_cache_row()
If the function receives an invalid user ID, it will issue an Undefined
offset notice, as it does not check for actual existence of the ID in
the users' cache after calling user_cache_array_rows().

PHPDoc does state that a valid user ID is expected, but since the
function includes error handling code, it makes sense to account for
this case, unlikely as it may be.

Fixes #25850