BugFilterQuery check for later modifications
The internal sql query is built and ready after the constructor method,
and regenerated after changing query type with set_query_type().
However, if externally adding clauses throguh add_join(), add_where(),
etc, those new parts are not translated to the final query
automatically.

Add a flag that keeps track of the modified state, so if the query is
executed after modifications, the internal sql is rebuilt.

Fixes: #24600