Fix warning in config_get_global()
When config_eval() references an unknown config option, the code throws
an ERROR_CONFIG_OPT_NOT_FOUND warning.

To avoid this unwanted behavior, we now check that the config option to
replace actually exists before calling config_get()/config_get_global().
If not, the variable will be left as-is in the returned value.

Fixes #26798