Fix VersionTest errors
Use of __construct() caused the following error:
array_merge(): Expected parameter 1 to be an array, null given

This is because the parent Constructor was not called, and so the
expected $data property was not set.

Overriding __construct() is not best practice anyway, the function was
renamed to setUp(), which fixes the problem.