Use parameterized string for confirmation messages
Previously, the confirmation message was built manually in several
places throughout the code, concatenating strings with variables.

We now use a string with parameters, which is fed to sprintf prior to
display. This gives translators more control to over the final message,
and allows removing now-unused $s_in_project string.

Updated strings:
  - $s_confirm_custom_field_deletion
  - $s_confirm_used_custom_field_deletion
  - $s_confirm_custom_field_unlinking
  - $s_config_delete_sure
  - $s_confirm_file_delete_msg
  - $s_delete_account_sure_msg
  - $s_query_delete_msg
  - $s_remove_user_sure_msg
  - $s_version_delete_sure

Issue #27779