Add resolved percentage on Roadmap section title
On the Roadmap page, a version's progress is now displayed on the right
of the section's title, with a tooltip showing the number of issues
resolved vs planned.

The resolved_progress language string has been modified to remove the
3rd placeholder (percentage) which is redundant.

Script has been refactored to use a new internal RoadmapProgress class,
simplifying code and removing some duplication.

Fixes #28182