Ignore E_DEPRECATED notices in PHP 8.1 builds
This is a hack to enable PHP 8.1 builds with PHPUnit 6.5, which is the
latest version that can be used with MantisBT 2.26's minimum PHP
requirement (7.0).

Modified composer.json to use the following custom MantisBT packages:
- phpunit/phpunit
- phpunit/php-code-coverage
- phpunit/php-file-iterator

The original repos were forked to the mantisbt organization, then a
branch was created off the latest tag; minimal code changes to avoid
deprecation notices in PHP 8.1 were committed.

We should be able to get rid of this when we update our PHP requirement
to 7.2, which will allow us to use PHPUnit 8.

Fixes #29882