Fix list.js scrolling regression patch
The list.js patch for the scrolling regression (issue #30494, commit
455eb76d2dc940af7f34448887b63e5b40ceefce) introduced another regression,
because preventing the default click event's execution not only applied
to the pagination buttons, but also to the checkboxes and the Apply
button as well, basically making the whole Manage Accounts section
unusable.

To fix the problem, we now only execute preventDefault() when the
pagination buttons are clicked.

Fixes #30494