Add canonical meta tag to view issue page
It is possible for issue URLs to have additional parameters in some 
situations. Search engines such as Google will index better if they 
know which are the set of canonical URLs. It reduces load through 
over-indexing, allows the search engine to prioritise the indexing of 
URLs that are correct, and avoids extraneous or non-ideal URLs getting 
into the index.

Fixes #31833, PR https://github.com/mantisbt/mantisbt/pull/1862