Fix FilterTest::testGetIssuesForUserForUnassignedNoTargetUser
The test case incorrectly failed when executed against a database
containing one or more unassigned sticky issues.

This is because filter_get_bug_rows() returns data sorted with sticky
issues first, and the test case made a bad assumption that the issue
created for the test case will be at index 0.