Remove unnecessary check on Version Id
It does not make any sense for prepare_version_string() to make a
strict-type comparison of the Version Id against false.

The same goes for the null Version Id check in version_full_name().

The Caller should take care of passing a valid Version. If it is not,
just throw an Exception.

Fixes #32831