Remove version_cache_row()'s 2nd parameter
$p_trigger_errors allows the function to conditionally throw an
Exception or return False when the Version can't be found.

This is a legacy construct going back to when trigger_error() was used
instead of Exceptions, and is only used once in version_exists().

To simplify code, the parameter is removed and a try/catch block is
used to replace the single use case.

Fixes #32832