Support setting enum configs via array of objects
In addition to support setting enum strings in the formation used in config_inc file,
The API now supports the same format that get config returns for enums,
which is an array of objects, each containing and id and a name property.
Note that the label field returns by get configs API, is not supported in set configs.

Fixes #32258