Add admin check to detect duplicate e-mails
When email_ensure_unique = ON, it is an error to have multiple user
accounts with the same email address in the database, which can happen
when this config's value is changed from OFF to ON.

This check facilitates identification of the offending accounts, helping
admins to fix them.

When email_ensure_unique = OFF, duplicates are shown as a warning, so
admins can anticipate problems when planning a switch to unique emails.

Fixes #32787