Use colspan=1 for date range filters
I'm not really sure why the date submitted and last updated fields were
set to span 2 columns in the first place. The browser will take care of
dynamically resizing the filter table when the UI is expanded to display
the date range selection fields, so in my opinion it is not necessary.

Switching these to single-column span frees some space, so new filter
fields can be easily inserted without having to completely reorder the
filter area (see issue #32726).